On power button, do not shut down, but prompt
This only became apparent after using dotfiles on a laptop, because with
that it's pretty common use the power button to initiate a shutdown, but
I still like to confirm it.

The default action to be chosen when the power button is pressed is
configured in the ksmserverrc instead (shutdown in this case).